# swagger-spec/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "patch": {
        "operationId": "UserController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Memperbarui profil dasar akun sendiri",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_removeMyAccount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Akun berhasil dianonimkan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menghapus akun sendiri (Soft Delete & Anonimisasi)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me/profile": {
      "patch": {
        "operationId": "UserController_updateMyProfileDetail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Memperbarui rincian profil detail pengguna (KTP, Bank, Alamat KTP)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me/pin": {
      "post": {
        "operationId": "UserController_setPin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Membuat atau memperbarui PIN keamanan (6 digit)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me/pin/verify": {
      "post": {
        "operationId": "UserController_verifyPin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Memverifikasi PIN keamanan untuk transaksi",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pengguna berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Payload tidak valid"
          },
          "409": {
            "description": "Email atau Nomor HP sudah terdaftar"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Membuat pengguna baru",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daftar pengguna berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan semua daftar pengguna",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me/avatar": {
      "post": {
        "operationId": "UserController_uploadAvatar",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mengunggah dan memperbarui avatar pengguna",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail pengguna ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan detail pengguna berdasarkan ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data pengguna berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Memperbarui data pengguna berdasarkan ID",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pengguna berhasil dihapus atau dianonimkan"
          },
          "404": {
            "description": "User tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menghapus pengguna berdasarkan ID (Hard/Soft Guard)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/status": {
      "patch": {
        "operationId": "UserController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mengubah status user (Active/Suspended/Blocked/Inactive)",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/reviews/users/{userId}": {
      "get": {
        "operationId": "ReviewsController_getUserRatingSummary",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ringkasan ulasan ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Melihat ringkasan dan daftar ulasan user",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/api/rewards/me": {
      "get": {
        "operationId": "RewardsController_getMyRewardSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ringkasan poin ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Melihat ringkasan saldo poin dan riwayat reward saya",
        "tags": [
          "Rewards"
        ]
      }
    },
    "/api/rewards/earn": {
      "post": {
        "operationId": "RewardsController_earnPoints",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EarnRewardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Poin berhasil ditambahkan"
          },
          "403": {
            "description": "Bukan user wilayah Anda"
          },
          "404": {
            "description": "User tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menambahkan poin reward ke user (Admin Only)",
        "tags": [
          "Rewards"
        ]
      }
    },
    "/api/rewards/redeem": {
      "post": {
        "operationId": "RewardsController_redeemPoints",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRewardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Poin berhasil ditukarkan"
          },
          "400": {
            "description": "Saldo poin tidak mencukupi"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menukarkan poin reward milik pengguna",
        "tags": [
          "Rewards"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nebeng App API Specification",
    "description": "Dokumentasi API Backend Platform Nebeng Transportasi & Pengiriman Paket (Escrow Wallet & Dual QR Checkpoint)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Masukkan token JWT: Bearer <token>",
        "in": "header"
      }
    },
    "schemas": {
      "UpdateUserMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jhons"
          },
          "email": {
            "type": "string",
            "example": "jhons@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "0812345678913"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "regionId": {
            "type": "string",
            "example": "1"
          },
          "avatar": {
            "type": "string",
            "example": "base64_string_or_url"
          }
        }
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1029103840"
          },
          "regionId": {
            "type": "object",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Jhons"
          },
          "email": {
            "type": "string",
            "example": "jhon@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "1029391032211"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "regional",
              "operator",
              "mitra",
              "customer"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "blocked",
              "inactive",
              "deleted"
            ]
          },
          "statusVerification": {
            "type": "string",
            "enum": [
              "unverified",
              "pending",
              "approved",
              "rejected"
            ]
          },
          "avatar": {
            "type": "object",
            "example": "https://example.com/avatar.jpg"
          },
          "rewardPoints": {
            "type": "number",
            "example": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "role",
          "status",
          "statusVerification",
          "rewardPoints",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserProfileDto": {
        "type": "object",
        "properties": {
          "ktpNumber": {
            "type": "string",
            "example": "3374123456789001",
            "description": "Nomor KTP 16 digit"
          },
          "fullNameKtp": {
            "type": "string",
            "example": "Budi Santoso",
            "description": "Nama lengkap sesuai KTP"
          },
          "addressKtp": {
            "type": "string",
            "example": "Jl. Merdeka No. 12, Jakarta",
            "description": "Alamat lengkap KTP"
          },
          "faceImageUrl": {
            "type": "string",
            "example": "https://storage.nebeng.com/faces/user-123.jpg",
            "description": "URL foto wajah / Face ID scan"
          },
          "bankName": {
            "type": "string",
            "example": "BCA",
            "description": "Nama Bank pencairan saldo"
          },
          "bankAccountNumber": {
            "type": "string",
            "example": "1234567890",
            "description": "Nomor Rekening Bank"
          },
          "bankAccountHolder": {
            "type": "string",
            "example": "Budi Santoso",
            "description": "Nama Pemilik Rekening Bank"
          }
        }
      },
      "SetPinDto": {
        "type": "object",
        "properties": {
          "pin": {
            "type": "string",
            "example": "123456",
            "description": "PIN 6 digit angka"
          }
        },
        "required": [
          "pin"
        ]
      },
      "VerifyPinDto": {
        "type": "object",
        "properties": {
          "pin": {
            "type": "string",
            "example": "123456",
            "description": "PIN 6 digit angka"
          }
        },
        "required": [
          "pin"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jhons"
          },
          "email": {
            "type": "string",
            "example": "jhons@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "0812345678913"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "regional",
              "operator",
              "mitra",
              "customer"
            ],
            "example": "customer"
          },
          "regionId": {
            "type": "string",
            "example": "1"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "blocked",
              "inactive",
              "deleted"
            ],
            "default": "active"
          },
          "avatar": {
            "type": "string",
            "example": "base64_string_or_url"
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "password",
          "role",
          "regionId"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jhons"
          },
          "email": {
            "type": "string",
            "example": "jhons@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "0812345678913"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "regional",
              "operator",
              "mitra",
              "customer"
            ],
            "example": "customer"
          },
          "regionId": {
            "type": "string",
            "example": "1"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "blocked",
              "inactive",
              "deleted"
            ],
            "default": "active"
          },
          "avatar": {
            "type": "string",
            "example": "base64_string_or_url"
          }
        }
      },
      "UpdateUserStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "blocked",
              "inactive",
              "deleted"
            ],
            "example": "suspended"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "EarnRewardDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID user penerima point",
            "example": "5"
          },
          "points": {
            "type": "number",
            "description": "jumlah point yang ditambahkan",
            "example": 100,
            "minimum": 1
          },
          "description": {
            "type": "string",
            "description": "keterangan sumber penamabahn point",
            "example": "Selesai transaksi nebeng penumpang #TRIP-13NA"
          }
        },
        "required": [
          "userId",
          "points"
        ]
      },
      "RedeemRewardDto": {
        "type": "object",
        "properties": {
          "points": {
            "type": "number",
            "description": "Jumlah poin yang ingin ditukarkan",
            "example": 50,
            "minimum": 1
          },
          "description": {
            "type": "string",
            "description": "keterangan penukaran point",
            "example": "Penukaran merchendase"
          }
        },
        "required": [
          "points"
        ]
      },
...
    }
  }
}